Refactor: Apply naming conventions to user module files
