+++ backend/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
-  testMatch: ["**/tests/**/*.test.ts", "**/__tests__/**/*.test.ts"],
+  testMatch: ["**/tests/**/*.spec.ts", "**/__tests__/**/*.spec.ts"],
 };
